[cap1203] working template with functional slider and mix mode
NOTE: in mix mode, the button behavior is implemented differently from pure
button mode:
1. Taps with duration longer than an pre-defined time is classified as cancelled
slide movement. The use case is that users may put their fingers on the bar but
then decide not to do the slide, thus it's a cancelling behavior.
2. Press state is sent to kscan listener only at the end of one update cycle.
While in pure button mode, the state change is sent immediately when an
interrput is received. This is a compromise, since to determine a slider
movement, the previous state is needed.
3. Slide movement is determined by the linear relative status pattern change.
4. Multi-touch is supported, but not recommended for reliable usage in daily
routine. The timing requirement is so strict and all presses and releases should
be performed at the same time to generate a single interrupt. Otherwise, it may
be recognized as a slide.
5. The algorithm design is from top to bottom, i.e. only the supported movements
are well recognized. If un-regular movements is performed, the output may not be
what users expected. For example, when initially tapping the 1st and 3rd pad,
and then touch the 2nd pad, there will be no output; on the other hand, if
pressing the 1st and 2nd pads first, then 3rd pad, this is recognized as slide
with delta 2/-2. However, this pattern is not within the design of algorithm.
And is not a useful movement anyway.